Reverse Proxy back to working
Crowdsec is removed for now. It has caused issues from day one. So, while I'd love to get it working I prefer stability for the time being. I'm not convinced it's actually a problem - but it IS noise.
Removing the Resource limitations for Caddy seemed to help, although there's no concrete evidence it was the problem.
I also altered the headers after much trial and error.  Many of these were superfluous - the result of pasting what I found from others. It's still not clear what the right levers to pull are here for security.