Maintenance/feb 2025 regularmaintenance (#114)
* Bump Caddy to 2.9.1

Along with arguments for easier reading

* Bump and pin Pihole/Unbound to 2024.07.0

Pinning keeps things a little more reliable

* Bump Socket Proxy to 0.3.0

Completes #105

* Bump Chrome to 4.28.1

* Minor variable change for Tandoor

* Minor cleanup

* Config folders and API Access

* CPU Min/Max

* Hygiene

* Bump Diun to 4.28

* Ditch Automatic Ripping Machine

Not planning to use this any more. If I do add a drive, there's alternatives (mkvtools)

* Cleanup Tandoor icons + Dashboard widget

* Better logging and minor error scenarios

This script is still not good enough; it does not catch port differences well.

* Reverse Proxy back to working

Crowdsec is removed for now. It has caused issues from day one. So, while I'd love to get it working I prefer stability for the time being. I'm not convinced it's actually a problem - but it IS noise.
Removing the Resource limitations for Caddy seemed to help, although there's no concrete evidence it was the problem.
I also altered the headers after much trial and error.  Many of these were superfluous - the result of pasting what I found from others. It's still not clear what the right levers to pull are here for security.

* Changes to Resources and Homepage config

Improving Tandoor's resource allocation seemed to improve load times.